feat: document title change and time function name change
A review of the FIG paper on the deformation model identified two items
that may be confusing to the documents audience.

* the title, Deformation Model Functional Model.  This is changed to
  "Functional Model of Crustal Deformation"

* the names "velocity" and "acceleration" for time functions.  These
  are changed to the more correct "linear" and "quadratic"

Also Figure 1 illustrating the functional model is revised to
hopefully provide a clearer picture of the relationship of the
components of the functional model, and to update the time function
name "velocity" to "linear".